chore: moved When_Source_Changed to a UI sample and more composition cleanup
